Upgrading to SF-4.1.0. Getting rid of \DependencyResolver\ as well as migrating to int[] data type for HashBins.

+++ src/SoundFingerprinting.Solr/Converters/IDictionaryToHashConverter.cs
@@ -4,8 +4,8 @@ namespace SoundFingerprinting.Solr.Converters
 
     internal interface IDictionaryToHashConverter
     {
-        Dictionary<int, long> FromHashesToSolrDictionary(long[] hashBins);
+        Dictionary<int, int> FromHashesToSolrDictionary(int[] hashBins);
 
-        long[] FromSolrDictionaryToHashes(IDictionary<int, long> hashTables);
+        int[] FromSolrDictionaryToHashes(IDictionary<int, int> hashTables);
     }
 }
